requirements.txt
asgiref==3.11.1
Authlib==1.6.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==6.0.3
djangorestframework==3.16.1
idna==3.11
Markdown==3.10.2
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
